+++ java/src/main/java/io/monohull/service/BundleService.java
@@ -48,8 +48,9 @@ public class BundleService {
         ImageConfigPayload icPayload = new ImageConfigPayload(
             ic.getClient(), ic.getProject(), ic.getMaximoVersion(),
             ic.getAppImage(), ic.getDbImage(), ic.getAdmImage(),
-            ic.getDbVendor(), ic.getDbName(), ic.getDbContainerPort(),
-            ic.getHostVolumePath(), ic.getDbVolumeName(), ic.getWorkspacePath(),
+            ic.getDbVendor(), ic.getDbName(), ic.getDbContainerPort(), ic.getDbCommand(),
+            ic.getHostVolumePath(), ic.getDbVolumeName(), ic.getDbVolumeTarget(),
+            ic.getWorkspacePath(),
             ic.getAppHttpPort(), ic.getAppHttpsPort(), ic.getDbPort(),
             ic.getMockHostPort(), ic.getSmtpHostPort(), ic.getSmtpUiHostPort(),
             ic.getDbExtraEnv(), ic.getDbExtraBinds(),
@@ -235,8 +236,10 @@ public class BundleService {
         ice.setDbVendor(ic.dbVendor());
         ice.setDbName(ic.dbName());
         ice.setDbContainerPort(ic.dbContainerPort());
+        ice.setDbCommand(ic.dbCommand());
         ice.setHostVolumePath(ic.hostVolumePath());
         ice.setDbVolumeName(ic.dbVolumeName());
+        ice.setDbVolumeTarget(ic.dbVolumeTarget());
         ice.setWorkspacePath(ic.workspacePath());
         ice.setAppHttpPort(ic.appHttpPort());
         ice.setAppHttpsPort(ic.appHttpsPort());